refactor(helpers): Reorganize and add helper to their own utility files
