+++ packages/test-case-component/src/helpers/index.ts
@@ -1,0 +1,10 @@
+export * from "./splitDocumentWithOffsets";
+export * from "./createDecorations";
+export * from "./getMarkDecorations";
+export * from "./getIdeFlashDecorations";
+export * from "./getSelections";
+export * from "./getThatMarks";
+export * from "./getSourceMarks";
+export * from "./addContentRangeDecorations";
+export * from "./typeGuards";
+export * from "./classesMap";
